+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * configuration cleanup
   * renamed net.xml to network.xml
   * renamed docroot type to docroot name
+  * renamed docroot docroot to docroot documents
   * merge alises into docroot
   
 * lock files now container user, machine and command being executed